requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
boto3==1.34.101
pyyaml==6.0.2
